cleeeeean stuff up and finish the refactor
