+++ src/seagulls/cli/_di_container.py
@@ -1,21 +1,15 @@
 from pathlib import Path
 
 from dependency_injector.containers import DeclarativeContainer
-from dependency_injector.providers import Singleton, Dependency
+from dependency_injector.providers import Dependency, Singleton
 from seagulls.assets import AssetManager
-from seagulls.attacks import WizardFireballFactory
-from seagulls.cli._launch2_command import Launch2Command, GameManager
-from seagulls.engine import GameWindowFactory, GameControls, GameSceneManager, GameClock
-from seagulls.player import PlayerSeagull
-from seagulls.scenes import SimpleScene
-from seagulls.scenes._storybook import StorybookScene
-from seagulls.ui import DebugHud
-from seagulls.wizards import SimpleWizardFactory
+from seagulls.engine import GameClock, GameControls
+from seagulls.prefabs import AsyncGameSessionManager
 
+from ._seagulls_command import SeagullsCommand
 from ._example_command import ExampleCommand
 from ._launch_command import LaunchCommand
 from ._framework import LoggingClient
-from ._seagulls_command import SeagullsCommand
 
 
 class SeagullsDiContainer(DeclarativeContainer):
@@ -27,61 +21,53 @@ class SeagullsDiContainer(DeclarativeContainer):
 
     _game_clock = Singleton(GameClock)
     _game_controls = Singleton(GameControls)
-    _window_factory = Singleton(GameWindowFactory)
     _asset_manager = Singleton(
         AssetManager,
         assets_path=Path("assets"),
     )
 
-    _scene_manager = Singleton(GameSceneManager)
+    # _debug_hud = Singleton(
+    #     DebugHud,
+    #     scene_manager=_scene_manager,
+    #     controls=_game_controls,
+    #     clock=_game_clock,
+    # )
 
-    _debug_hud = Singleton(
-        DebugHud,
-        scene_manager=_scene_manager,
-        controls=_game_controls,
-        clock=_game_clock,
-    )
+    # _fireball_factory = Singleton(
+    #     WizardFireballFactory,
+    #     clock=_game_clock,
+    #     scene_manager=_scene_manager,
+    #     asset_manager=_asset_manager,
+    # )
 
-    _fireball_factory = Singleton(
-        WizardFireballFactory,
-        clock=_game_clock,
-        scene_manager=_scene_manager,
-        asset_manager=_asset_manager,
-    )
+    # _wizard_factory = Singleton(
+    #     SimpleWizardFactory,
+    #     scene_manager=_scene_manager,
+    #     fireball_factory=_fireball_factory,
+    #     clock=_game_clock,
+    #     asset_manager=_asset_manager,
+    # )
+    # _player_seagull = Singleton(
+    #     PlayerSeagull,
+    #     controls=_game_controls,
+    #     clock=_game_clock,
+    #     scene_manager=_scene_manager,
+    #     asset_manager=_asset_manager,
+    # )
+    # _simple_scene = Singleton(
+    #     SimpleScene,
+    #     player=_player_seagull,
+    #     asset_manager=_asset_manager,
+    #     wizard_factory=_wizard_factory,
+    #     debug_hud=_debug_hud,
+    # )
 
-    _wizard_factory = Singleton(
-        SimpleWizardFactory,
-        scene_manager=_scene_manager,
-        fireball_factory=_fireball_factory,
-        clock=_game_clock,
-        asset_manager=_asset_manager,
-    )
-    _player_seagull = Singleton(
-        PlayerSeagull,
-        controls=_game_controls,
-        clock=_game_clock,
-        scene_manager=_scene_manager,
-        asset_manager=_asset_manager,
-    )
-    _simple_scene = Singleton(
-        SimpleScene,
-        player=_player_seagull,
-        asset_manager=_asset_manager,
-        wizard_factory=_wizard_factory,
-        debug_hud=_debug_hud,
-    )
-    _storybook_scene = Singleton(StorybookScene)
+    _game_session_manager = Singleton(AsyncGameSessionManager)
 
     root_command = Singleton(SeagullsCommand)
     launch_command = Singleton(
         LaunchCommand,
-        window_factory=_window_factory,
-        scene=_storybook_scene,
-        scene_manager=_scene_manager,
-        clock=_game_clock,
-        controls=_game_controls,
+        game_session_manager=_game_session_manager,
     )
-    game_manager = Singleton(GameManager)
-    launch2_command = Singleton(Launch2Command, game_manager=game_manager)
 
     example_command = Singleton(ExampleCommand)
