+++ src/seagulls/creatures/__init__.py
@@ -1,0 +1,5 @@
+from ._bird import Bird
+
+__all__ = [
+    "Bird",
+]
