+++ src/seagulls/engine/__init__.py
@@ -1,18 +1,11 @@
-from ._game_clock import (
-    GameTimeUpdater,
-    GameTimeProvider,
-    GameClock,
-)
+from ._game_clock import GameClock
 from ._game_controls import GameControls
-from ._game_window import (
-    GameWindow,
-    GameWindowFactory,
-)
-from ._game_scene import (
-    GameScene,
-    GameSceneObjects,
-    GameSceneManager,
-)
+from ._game_scene import IGameScene
+from ._surface_renderer import SurfaceRenderer
+from ._game_scene_manager import IProvideGameScenes
+from ._game_session import IGameSession
+from ._game_session_manager import IProvideGameSessions
+
 from ._pyagme import (
     Rect,
     Surface,
@@ -21,30 +14,28 @@ from ._pyagme import (
     Vector2,
     Vector3,
 )
-from ._game_object import GameObject
+from ._game_object import (
+    GameObject,
+    GameObjectsCollection
+)
 from ._collisions import (
     flag_from_string,
     CollidableObject
 )
 from ._game_settings import GameSettings
-from ._scenedb import (
-    ILocateScenes,
-    ScenesClient
-)
 
 __all__ = [
     "flag_from_string",
     "CollidableObject",
-    "GameTimeUpdater",
-    "GameTimeProvider",
+    "IGameScene",
+    "IProvideGameScenes",
+    "IProvideGameSessions",
+    "IGameSession",
+    "SurfaceRenderer",
     "GameClock",
     "GameControls",
     "GameObject",
-    "GameWindow",
-    "GameWindowFactory",
-    "GameScene",
-    "GameSceneObjects",
-    "GameSceneManager",
+    "GameObjectsCollection",
     "GameSettings",
     "Rect",
     "Surface",
@@ -52,6 +43,4 @@ __all__ = [
     "PixelArray",
     "Vector2",
     "Vector3",
-    "ILocateScenes",
-    "ScenesClient",
 ]
