+++ src/seagulls/engine/_game_object.py
@@ -1,4 +1,5 @@
-from abc import abstractmethod, ABC
+from abc import ABC, abstractmethod
+from typing import List, Callable
 
 from ._pyagme import Surface
 
@@ -6,13 +7,23 @@ from ._pyagme import Surface
 class GameObject(ABC):
 
     @abstractmethod
-    def update(self) -> None:
+    def tick(self) -> None:
         pass
 
     @abstractmethod
     def render(self, surface: Surface) -> None:
         pass
 
-    @abstractmethod
-    def is_destroyed(self) -> bool:
-        pass
+
+class GameObjectsCollection:
+    _game_objects: List[GameObject]
+
+    def __init__(self) -> None:
+        self._game_objects = []
+
+    def add(self, game_object: GameObject) -> None:
+        self._game_objects.append(game_object)
+
+    def apply(self, func: Callable[[GameObject], None]) -> None:
+        for game_object in self._game_objects:
+            func(game_object)
