+++ src/seagulls/engine/_game_scene_manager.py
@@ -1,0 +1,10 @@
+from abc import ABC, abstractmethod
+
+from ._game_scene import IGameScene
+
+
+class IProvideGameScenes(ABC):
+
+    @abstractmethod
+    def get_scene(self, name: str) -> IGameScene:
+        pass
