+++ src/seagulls/engine/_game_session.py
@@ -1,0 +1,16 @@
+from abc import abstractmethod, ABC
+
+
+class IGameSession(ABC):
+
+    @abstractmethod
+    def start(self) -> None:
+        pass
+
+    @abstractmethod
+    def wait_for_completion(self) -> None:
+        pass
+
+    @abstractmethod
+    def stop(self) -> None:
+        pass
