+++ src/seagulls/prefabs/__init__.py
@@ -1,0 +1,13 @@
+from ._background import SeagullsBackground
+from ._game_scene import SeagullGameScene
+from ._game_scene_manager import SeagullsGameSceneManager
+from ._game_session import AsyncGameSession
+from ._game_session_manager import AsyncGameSessionManager
+
+__all__ = [
+    "SeagullsBackground",
+    "SeagullGameScene",
+    "SeagullsGameSceneManager",
+    "AsyncGameSession",
+    "AsyncGameSessionManager",
+]
